+++ packages/subgraph/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `Granted` & `Revoked` event in `PluginRepo`
-- Add `one-to-many` relationship from `Dao` & `PluginRepo` to `Permission`
+- Add one-to-many relationship from `Dao` & `PluginRepo` to `Permission`
 
 ### Changed